Automatic conversion from a Python sequence into Collection<foo>
When defining collections with OTTypedCollectionInterfaceObjectHelper,
struct traitsPythonType must be defined like this:
  template <> struct traitsPythonType<OT::Interface>
  {
    typedef _PyObject_ Type;
  };
and the following two functions must be defined:

  template <> inline bool canConvert< _PyObject_, OT::Interface>(PyObject * pyObj)
  template <> inline OT::Interface convert< _PyObject_, OT::Interface>(PyObject * pyObj)

If Collection<foo> is used in a different module from where it is defined, this
module will fail to build:
  error: invalid use of incomplete type ‘struct OT::traitsPythonType<foo>’

These functions must then be written into a separate file.  It had been decided to
write them into <module>Collection.i, where <module> is named after <module>TemplateDefs.i.
Then
  %include <module>Collection.i
is written into foo.i, and the same line is inserted into xxx_module.i if this module
needs these functions in order to be compiled.